back-end/index: seperates the express app from the index file
Seperates the express app from the index.js file for testing and
seperation of duty purposes. With this commit, index.js becomes our
startup code, while app becomes our express initialization code. In
app.js, only the express app object is exported. This means that we can
test the app object without ever having to start the server, which makes
testing routes way easier.
